requirements.txt
anyio==3.7.0
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.98.0
google-api-core==2.11.1
google-auth==2.20.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
googleapis-common-protos==1.59.1
graphqlclient==0.2.4
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
ipdb==0.13.13
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
orjson==3.9.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.23.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
Pygments==2.15.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.6.3
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
